Link in Bio Banner: Address review comments
