--- client/sites-dashboard/components/link-in-bio-banner/BannerType.ts
@@ -1,6 +1,0 @@
-export enum BannerType {
-	None,
-	Tile,
-	DoubleTile,
-	Row,
-}
