+++ client/sites-dashboard/components/sites-grid.tsx
@@ -1,7 +1,7 @@
 import { css } from '@emotion/css';
 import classnames from 'classnames';
 import { MEDIA_QUERIES } from '../utils';
-import { useLinkInBioBanner } from './link-in-bio-banner/use-link-in-bio-banner';
+import { LinkInBioBanner } from './link-in-bio-banner/link-in-bio-banner';
 import { SitesGridItem } from './sites-grid-item';
 import { SitesGridItemLoading } from './sites-grid-item-loading';
 import type { SiteExcerptData } from 'calypso/data/sites/site-excerpt-types';
@@ -35,8 +35,6 @@ export const SitesGrid = ( { sites, isLoading, className }: SitesGridProps ) => 
 	if ( params.get( 'sitecount' ) ) {
 		sites = sites.slice( 0, Number( params.get( 'sitecount' ) ) );
 	}
-	const { getBanner } = useLinkInBioBanner();
-	const banner = getBanner( sites, 'grid' );
 	return (
 		<div className={ classnames( container, className ) }>
 			{ isLoading
@@ -44,7 +42,7 @@ export const SitesGrid = ( { sites, isLoading, className }: SitesGridProps ) => 
 						.fill( null )
 						.map( ( _, i ) => <SitesGridItemLoading key={ i } delayMS={ i * 150 } /> )
 				: sites.map( ( site ) => <SitesGridItem site={ site } key={ site.ID } /> ) }
-			{ banner }
+			<LinkInBioBanner siteCount={ sites.length } displayMode={ 'grid' } />
 		</div>
 	);
 };
